move email server action to route handler
Couldn't get my server action <10s max execution time. 😭
